hf_deploy_requirements.txt
gradio==6.12.0
numpy==2.4.4
pandas==3.0.2
torch==2.11.0+cpu
torch_geometric==2.7.0
tqdm==4.67.3
transformers==5.5.4
llama-cpp-python